Remove session set
This session setter saves in Redis a set of session ids as a history.
We have not seen any other functionality for this data set.

Due to the memory problems caused in Redis, we have decided to
eliminate this history since, due to Redis limitations, it is not
possible to apply a TTL and expire it automatically.

SNS-1575